requirements.txt
# AI Trading Assistant — 의존성 패키지
# Python 3.11+

# 시장 데이터
yfinance==0.2.54
pandas==2.2.3
numpy==2.1.3

# OCR
pytesseract==0.3.13
Pillow==11.1.0

# 대시보드
streamlit==1.43.2
plotly==5.24.1

# 텔레그램 봇
python-telegram-bot==21.10

# 설정 파일
PyYAML==6.0.2

# 유틸리티
python-dateutil==2.9.0
requests==2.32.3
pytz==2025.1

# 테스트
pytest==8.3.5
pytest-asyncio==0.25.3
